Move optional fields into a separate component
